Add SuspendingCloseable
